+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Set the route title for the asset and manage plugins pages
 - Change the archived asset name icon to use a unicode icon that can also be rendered as the `document.title`
 - Use the entity name component to render asset names in the asset selector component
+- Use the entity name component to render asset names in the asset hierarchy dialog
 
 ## [0.6.2] - 2022-09-03